wip on data fetching repos. Plus some refactoring work
